Use fixed 32bit sign argument to error functions
Ideally sign should be bool, but we don't want to depend on stdbool.h
and sometimes (e.g. in powf) it's more efficient to pass a non-zero
value than 1 to indicate that the sign should be set.

While working on double precision versions i realized that the unsigned
long argument is fragile (may fail on 32bit targets only).

+++ math/math_config.h
@@ -138,10 +138,10 @@ issignalingf_inline (float x)
 # define NOINLINE
 #endif
 
-HIDDEN float __math_oflowf (unsigned long);
-HIDDEN float __math_uflowf (unsigned long);
-HIDDEN float __math_may_uflowf (unsigned long);
-HIDDEN float __math_divzerof (unsigned long);
+HIDDEN float __math_oflowf (uint32_t);
+HIDDEN float __math_uflowf (uint32_t);
+HIDDEN float __math_may_uflowf (uint32_t);
+HIDDEN float __math_divzerof (uint32_t);
 HIDDEN float __math_invalidf (float);
 
 /* Shared between expf, exp2f and powf.  */
